fixup! Add wallet adapter for NuFi
